Refactor statistics components
