+++ frontend/src/app/pages/exercises/exercise/shared/exercise-statistics/area-statistics.service.ts
@@ -3,8 +3,9 @@ import { Store } from '@ngrx/store';
 import type { UUID } from 'digital-fuesim-manv-shared';
 import type { AreaStatistics } from 'digital-fuesim-manv-shared/dist/models/utils/area-statistics';
 import type { Observable } from 'rxjs';
-import { BehaviorSubject, switchMap } from 'rxjs';
+import { BehaviorSubject, map, switchMap } from 'rxjs';
 import type { AppState } from 'src/app/state/app.state';
+import { decimateStatistics } from './shared/decimate-statistics';
 
 @Injectable({
     providedIn: 'root',
@@ -20,7 +21,7 @@ export class AreaStatisticsService {
     /**
      * Emits the statistics for the current area and the accompanying correct exerciseTime
      */
-    public areaStatistics$: Observable<AreaStatisticsEntry[]> =
+    public readonly areaStatistics$: Observable<AreaStatisticsEntry[]> =
         this.areaId$.pipe(
             switchMap((areaId) =>
                 this.store.select((state) =>
@@ -37,6 +38,13 @@ export class AreaStatisticsService {
                 )
             )
         );
+
+    /**
+     * The {@link areaStatistics$} decimated to reduce the amount of data per emit
+     */
+    public readonly decimatedAreaStatistics$ = this.areaStatistics$.pipe(
+        map((statistics) => decimateStatistics(statistics))
+    );
 
     public setAreaId(areaId: UUID | null) {
         this.areaId = areaId;
