+++ frontend/src/app/pages/exercises/exercise/shared/exercise-statistics/exercise-statistics.module.ts
@@ -3,16 +3,12 @@ import { CommonModule } from '@angular/common';
 import { NgbDropdownModule } from '@ng-bootstrap/ng-bootstrap';
 import { SharedModule } from 'src/app/shared/shared.module';
 import { ExerciseStatisticsModalComponent } from './exercise-statistics-modal/exercise-statistics-modal.component';
-import { PatientStatisticsComponent } from './patient-statistics/patient-statistics.component';
-import { VehicleStatisticsComponent } from './vehicle-statistics/vehicle-statistics.component';
-import { PersonnelStatisticsComponent } from './personel-statistics/personnel-statistics.component';
+import { TimeLineAreaChartComponent } from './time-line-area-chart/time-line-area-chart.component';
 
 @NgModule({
     declarations: [
         ExerciseStatisticsModalComponent,
-        PatientStatisticsComponent,
-        VehicleStatisticsComponent,
-        PersonnelStatisticsComponent,
+        TimeLineAreaChartComponent,
     ],
     imports: [CommonModule, NgbDropdownModule, SharedModule],
 })
