--- frontend/src/app/pages/exercises/exercise/shared/exercise-statistics/patient-statistics/patient-statistics.component.html
@@ -1,2 +1,0 @@
-<h3>Patienten</h3>
-<canvas #chart class="w-100"></canvas>
