--- frontend/src/app/pages/exercises/exercise/shared/exercise-statistics/personel-statistics/personnel-statistics.component.html
@@ -1,5 +1,0 @@
-<h3>
-    Einsatzkräfte
-    <small class="text-muted">(weder im Transfer noch im Fahrzeug)</small>
-</h3>
-<canvas #chart class="w-100"></canvas>
