+++ frontend/src/app/pages/exercises/exercise/shared/exercise-statistics/time-line-area-chart/time-line-area-chart.component.html
@@ -1,2 +1,1 @@
-<h3>Fahrzeuge</h3>
 <canvas #chart class="w-100"></canvas>
